# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateDonationRequest:
    properties:
      donation_date:
        type: string
      event_id:
        type: string
      location_id:
        type: string
      name:
        type: string
      status:
        enum:
        - selesai
        - batal
        - pending
        type: string
    required:
    - donation_date
    - location_id
    - name
    - status
    type: object
  dto.ErrorWrapper:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.SuccessWrapper:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.UpdateDonationRequest:
    properties:
      status:
        enum:
        - selesai
        - batal
        - pending
        type: string
    required:
    - status
    type: object
...
info:
  contact:
    email: support@donor-darah.duckdns.org
    name: Ulil Amry
    url: https://donor-darah.duckdns.org
  description: API untuk mengelola donor darah, stok darah, dan informasi pendonor
    dengan autentikasi OAuth.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://donor-darah.duckdns.org/terms
  title: Donor App API
  version: "1.0"
paths:
...
  /donations:
    get:
      description: Mengambil daftar semua data donasi dengan paginasi
      parameters:
      - default: 1
        description: Nomor halaman
        in: query
        name: page
        type: integer
      - default: 10
        description: Jumlah item per halaman
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil daftar donasi
          schema:
            $ref: '#/definitions/dto.SuccessWrapper'
        "500":
          description: Terjadi kesalahan internal
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
      security:
      - BearerAuth: []
      summary: Get all donations
      tags:
      - Donations
    post:
      consumes:
      - application/json
      description: Menambahkan data donasi baru ke sistem
      parameters:
      - description: Data Donasi Baru
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDonationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Donasi berhasil dibuat
          schema:
            $ref: '#/definitions/dto.SuccessWrapper'
        "400":
          description: Request tidak valid
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
        "500":
          description: Terjadi kesalahan internal
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
      security:
      - BearerAuth: []
      summary: Create a new donation
      tags:
      - Donations
  /donations/{id}:
    delete:
      description: Menghapus data donasi berdasarkan ID
      parameters:
      - description: ID Donasi
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Donasi berhasil dihapus
          schema:
            $ref: '#/definitions/dto.SuccessWrapper'
        "400":
          description: Format ID tidak valid
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
        "500":
          description: Terjadi kesalahan internal
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
      security:
      - BearerAuth: []
      summary: Delete a donation
      tags:
      - Donations
    get:
      description: Mengambil satu data donasi berdasarkan ID
      parameters:
      - description: ID Donasi
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil data donasi
          schema:
            $ref: '#/definitions/dto.SuccessWrapper'
        "400":
          description: Format ID tidak valid
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
        "404":
          description: Data tidak ditemukan
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
      security:
      - BearerAuth: []
      summary: Get donation by ID
      tags:
      - Donations
    put:
      consumes:
      - application/json
      description: Memperbarui data donasi yang sudah ada berdasarkan ID
      parameters:
      - description: ID Donasi
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Data Donasi yang Diperbarui
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateDonationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Donasi berhasil diperbarui
          schema:
            $ref: '#/definitions/dto.SuccessWrapper'
        "400":
          description: Format ID atau request tidak valid
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
        "500":
          description: Terjadi kesalahan internal
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
      security:
      - BearerAuth: []
      summary: Update a donation
      tags:
      - Donations
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"